mobile/android/components/extensions/test/mochitest/test_ext_pageAction.html

<!DOCTYPE HTML>
<html>
<head>
  <title>PageAction Test</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/SpawnTask.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/ExtensionTestUtils.js"></script>
  <script type="text/javascript" src="head.js"></script>
  <link rel="stylesheet" href="chrome://mochikit/contents/tests/SimpleTest/test.css"/>
</head>
<body>

<script type="text/javascript">
"use strict";

var {PageActions} = SpecialPowers.Cu.import("resource://gre/modules/PageActions.jsm", {});

let dataURI = "iVBORw0KGgoAAAANSUhEUgAAACQAAAAkCAYAAADhAJiYAAAC4klEQVRYhdWXLWzbQBSADQtDAwsHC1tUhUxqfL67lk2tdn+OJg0ODU0rLByqgqINBY6tmlbn7LMTJ5FaFVVBk1G0oUGjG2jT2Y7jxmmcbU/6iJ+f36fz+e5sGP9riCGm9hB37RG+scd4Yo/wsDXCZyIE2xuXsce4bY+wXkAsQtzYmExrfFgvkJkRbkzo1ehoxx5iXcgI/9iYUGt8WH9MqDXEcmNChmEYrRCf2SHWeYgQx3x0tLNRIeKQLTtEFyJEep4NTuhk8BC+yMrwEE3+iozo42d8gK7FAOkMsRiiN8QhW2ttSK5QTfRRV4QoymVeJMvPvDp7gCZigD613MN6yRFA3SWarow9QB9LCfG+NeF9qCtjAKOSQjCqVKhfVsiHEQ+grgx/lRGqUihAc1uL8EFD+KCRO+GrF4J61phcoRoPoEzkYhZYpykh5sMb7kOdIeY+jHKur4QI4Feh4AFX1nVeLxrAvQchGsBz5ls6wa2QdwcvIcE2863bTH79KOvsz/uUYJsp+J0pSzNlDckVqqVGUAF+n6uS7txcOl6wot4JVy70ufDLy4pWLUQVPE81pRI0mGe9oxLMHSeohHvMs/STUNaUK6vDPCvOyxMFDx4achehRDJmHnydnkPww5OFfLxrGIZBFDyYl4LpMzlTQFIP6AQx86w2UeYBccFpJrcKv5L9eGDtUAU6RIELqsB74uynjy/UBRF1gS5BTFxwQT1wTiXoUg9MH7m/3NZRRoi5IJytUbMgzv4Wc832+oQkiKgEehmyMkkpKsFkQV11QsRJL5rJYBLItQgRaUZEmnoZXsomz3vGiWw+I9KMF9SVFOqZEemZekli1jN3U/UOqhHHvC6oWWGElhfSpGdOk6+O9prdwvtLj5BjRsQxdRnot+Zeifpy/2/0stktKTRNLmbk0mwXyl8253fyojj+8rxOHNAhjjm5n0/5OOCGOKBzkrMO0Z75lvSAzKlrF32Z/3z8BqLAn+yMV7VhAAAAAElFTkSuQmCC";

let image = atob(dataURI);
const IMAGE_ARRAYBUFFER = Uint8Array.from(image, byte => byte.charCodeAt(0)).buffer;

async function background() {
  browser.test.assertTrue("pageAction" in browser, "Namespace 'pageAction' exists in browser");
  browser.test.assertTrue("show" in browser.pageAction, "API method 'show' exists in browser.pageAction");

  const tabs = await browser.tabs.query({active: true, currentWindow: true});

  browser.test.onMessage.addListener(msg => {
    if (msg === "pageAction-show") {
      browser.pageAction.show(tabs[0].id).then(() => {
        browser.test.sendMessage("page-action-shown");
      });
    } else if (msg === "pageAction-hide") {
      browser.pageAction.hide(tabs[0].id).then(() => {
        browser.test.sendMessage("page-action-hidden");
      });
    }
  });

  browser.pageAction.onClicked.addListener(tab => {
    browser.test.sendMessage("page-action-clicked", tab);
  });

  browser.test.sendMessage("ready", tabs[0]);
}

add_task(function* test_pageAction() {
  const extension = ExtensionTestUtils.loadExtension({
    background,
    manifest: {
      "name": "PageAction Extension",
      "page_action": {
        "default_title": "Page Action",
        "default_icon": {
          "18": "extension.png",
        },
      },
      "applications": {
        "gecko": {
          "id": "foo@bar.com",
        },
      },
    },
    files: {
      "extension.png": IMAGE_ARRAYBUFFER,
    },
  });

  yield extension.startup();
  const tab = yield extension.awaitMessage("ready");
  const uuid = `{${extension.uuid}}`;

  extension.sendMessage("pageAction-show");
  yield extension.awaitMessage("page-action-shown");
  ok(PageActions.isShown(uuid), "The PageAction should be shown");

  extension.sendMessage("pageAction-hide");
  yield extension.awaitMessage("page-action-hidden");
  ok(!PageActions.isShown(uuid), "The PageAction should be hidden");

  extension.sendMessage("pageAction-show");
  yield extension.awaitMessage("page-action-shown");
  ok(PageActions.isShown(uuid), "The PageAction should be shown");

  PageActions.synthesizeClick(uuid);
  const clickedTab = yield extension.awaitMessage("page-action-clicked");
  ok(PageActions.isShown(uuid), "The PageAction should still be shown after being clicked");
  is(clickedTab.id, tab.id, "Got the expected tab id in the pageAction.onClicked event");

  yield extension.unload();
  ok(!PageActions.isShown(uuid), "The PageAction should be removed after unload");
});
</script>

</body>
</html>
